refactor(onboarding): modularize onboarding into a feature-based architecture
- Extract onboarding logic from app/onboarding to features/onboarding
- Create reusable OnboardingModal component with customizable close button support
- Centralize step metadata and options into a dedicated config file
- Componentize StepHeader and StepNavigation for better maintainability
- Standardize Zod schema and TypeScript types into the feature directory
- Drastically simplify OnboardingPage to focus strictly on orchestration
