+++ features/onboarding/components/step-navigation.tsx
@@ -1,0 +1,26 @@
+import { StepperNext, StepperPrevious, useStepper } from '@/components/primitives/stepper';
+import { Button } from '@/components/ui/button';
+import { ArrowLeft, ArrowRight, Check } from 'lucide-react';
+import * as React from 'react';
+
+export function StepNavigation({ isFormValid }: { isFormValid: boolean }) {
+  const { isFirst, isLast, currentStep } = useStepper();
+
+  return (
+    <div className="flex justify-between">
+      <StepperPrevious asChild>
+        <Button variant="outline" className={isFirst ? 'invisible' : 'h-12 text-lg rounded-xl'}>
+          <ArrowLeft />
+          Back
+        </Button>
+      </StepperPrevious>
+
+      <StepperNext asChild disabled={currentStep === 'space-name' && !isFormValid}>
+        <Button className={'h-12 text-lg rounded-xl'}>
+          {isLast ? 'Complete Setup' : currentStep === 'invite' ? 'Skip' : 'Next'}
+          {isLast ? <Check /> : <ArrowRight />}
+        </Button>
+      </StepperNext>
+    </div>
+  );
+}
