+++ features/onboarding/types/request.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+import { onboardingSchema } from '@/features/onboarding/schemas/onboarding.schema';
+
+export type OnboardingValues = z.infer<typeof onboardingSchema>;
